refactor(app): refactor app and main
